extracts view logic from root page
