+++ test/admin/components/views/CustomDefault/index.tsx
@@ -7,6 +7,8 @@ import React from 'react'
 
 const Link = (LinkImport.default || LinkImport) as unknown as typeof LinkImport.default
 
+import { customViewPath } from 'test/admin/shared.js'
+
 import './index.scss'
 
 const baseClass = 'custom-default-view'
@@ -53,9 +55,18 @@ export const CustomDefaultView: React.FC<AdminViewProps> = ({ initPageResult }) 
           Here is a custom admin view that was added in the Payload config. It uses the Default
           Template, so the sidebar is rendered.
         </p>
-        <div className={`${baseClass}__controls`}>
+        <div className="custom-view__controls">
           <Button Link={Link} buttonStyle="secondary" el="link" to={`${adminRoute}`}>
             Go to Dashboard
+          </Button>
+          &nbsp; &nbsp; &nbsp;
+          <Button
+            Link={Link}
+            buttonStyle="secondary"
+            el="link"
+            to={`${adminRoute}/${customViewPath}`}
+          >
+            Go to Custom View
           </Button>
         </div>
       </div>
