+++ test/admin/components/views/CustomMinimal/index.tsx
@@ -14,6 +14,8 @@ const Link = (LinkImport.default || LinkImport) as unknown as typeof LinkImport.
 
 import type { AdminViewProps } from 'payload/types.js'
 
+import { customViewPath } from 'test/admin/shared.js'
+
 import './index.scss'
 
 const baseClass = 'custom-minimal-view'
@@ -35,17 +37,20 @@ export const CustomMinimalView: React.FC<AdminViewProps> = ({ initPageResult }) 
         <h1>Custom Admin View</h1>
         <p>Here is a custom admin view that was added in the Payload config.</p>
         <div className={`${baseClass}__controls`}>
-          <Button
-            Link={Link}
-            className={`${baseClass}__login-btn`}
-            el="link"
-            to={`${adminRoute}/login`}
-          >
-            Go to Login
-          </Button>
-          <Button Link={Link} buttonStyle="secondary" el="link" to={`${adminRoute}`}>
-            Go to Dashboard
-          </Button>
+          <div className="custom-view__controls">
+            <Button Link={Link} buttonStyle="secondary" el="link" to={`${adminRoute}`}>
+              Go to Dashboard
+            </Button>
+            &nbsp; &nbsp; &nbsp;
+            <Button
+              Link={Link}
+              buttonStyle="secondary"
+              el="link"
+              to={`${adminRoute}/${customViewPath}`}
+            >
+              Go to Custom View
+            </Button>
+          </div>
         </div>
       </div>
     </MinimalTemplate>
