+++ test/admin/config.ts
@@ -72,22 +72,22 @@ export default buildConfigWithDefaults({
     CollectionNoApiView,
     CustomViews1,
     CustomViews2,
-    // CollectionGroup1A,
-    // CollectionGroup1B,
-    // CollectionGroup2A,
-    // CollectionGroup2B,
-    // Geo,
-    // CustomIdTab,
-    // CustomIdRow,
+    CollectionGroup1A,
+    CollectionGroup1B,
+    CollectionGroup2A,
+    CollectionGroup2B,
+    Geo,
+    CustomIdTab,
+    CustomIdRow,
   ],
   globals: [
     GlobalHidden,
     GlobalNoApiView,
-    // Global,
+    Global,
     CustomGlobalViews1,
     CustomGlobalViews2,
-    // GlobalGroup1A,
-    // GlobalGroup1B,
+    GlobalGroup1A,
+    GlobalGroup1B,
   ],
   i18n: {
     translations: {
