+++ test/admin/globals/Group1A.ts
@@ -4,7 +4,6 @@ import { group1GlobalSlug } from '../slugs.js'
 
 export const GlobalGroup1A: GlobalConfig = {
   slug: group1GlobalSlug,
-  label: 'Group Globals 1',
   admin: {
     group: 'Group',
   },
@@ -14,4 +13,5 @@ export const GlobalGroup1A: GlobalConfig = {
       type: 'text',
     },
   ],
+  label: 'Group Globals 1',
 }
